requirements.txt
h5py==2.9.0
requests==2.21.0
runway-python
tensorflow-gpu==1.15.0
scipy
